+++ packages/gov-ui-kit/src/core/components/accordion/accordionContainer/accordionContainer.stories.tsx
@@ -1,4 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
+import { type RefAttributes } from 'react';
 import { Accordion } from '..';
 import { AccordionContainer, type IAccordionContainerProps } from './accordionContainer';
 
@@ -19,27 +20,28 @@ const meta: Meta<typeof AccordionContainer> = {
 
 type Story = StoryObj<typeof AccordionContainer>;
 
-const reusableStoryComponent = (props: IAccordionContainerProps, count: number) => (
-    <Accordion.Container {...props}>
-        {Array.from({ length: count }, (_, index) => (
-            <Accordion.Item key={`item-${index + 1}`} value={`item-${index + 1}`}>
-                <Accordion.ItemHeader>Item {index + 1} Header</Accordion.ItemHeader>
-                <Accordion.ItemContent>
-                    <div className="flex h-24 w-full items-center justify-center border border-dashed border-info-300 bg-info-100">
-                        Item {index + 1} Content
-                    </div>
-                </Accordion.ItemContent>
-            </Accordion.Item>
-        ))}
-    </Accordion.Container>
-);
-
+const reusableStoryComponent = (props: IAccordionContainerProps & RefAttributes<HTMLDivElement>, count: number) => {
+    return (
+        <Accordion.Container {...props}>
+            {Array.from({ length: count }, (_, index) => (
+                <Accordion.Item key={`item-${index + 1}`} value={`item-${index + 1}`}>
+                    <Accordion.ItemHeader>Item {index + 1} Header</Accordion.ItemHeader>
+                    <Accordion.ItemContent>
+                        <div className="flex h-24 w-full items-center justify-center border border-dashed border-info-300 bg-info-100">
+                            Item {index + 1} Content
+                        </div>
+                    </Accordion.ItemContent>
+                </Accordion.Item>
+            ))}
+        </Accordion.Container>
+    );
+};
 /**
  * Default usage example of a full Accordion component.
  */
 export const Default: Story = {
     args: { isMulti: false },
-    render: (args) => reusableStoryComponent(args, 1),
+    render: (args) => reusableStoryComponent(args as IAccordionContainerProps & RefAttributes<HTMLDivElement>, 1),
 };
 
 /**
@@ -47,7 +49,7 @@ export const Default: Story = {
  */
 export const SingleTypeItems: Story = {
     args: { isMulti: false },
-    render: (args) => reusableStoryComponent(args, 3),
+    render: (args) => reusableStoryComponent(args as IAccordionContainerProps & RefAttributes<HTMLDivElement>, 3),
 };
 
 /**
@@ -55,7 +57,7 @@ export const SingleTypeItems: Story = {
  */
 export const MultipleTypeItems: Story = {
     args: { isMulti: true, defaultValue: ['item-2', 'item-3'] },
-    render: (args) => reusableStoryComponent(args, 3),
+    render: (args) => reusableStoryComponent(args as IAccordionContainerProps & RefAttributes<HTMLDivElement>, 3),
 };
 
 export default meta;