+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `Accordion` core component
+-   Implement `DocumentParser` and `Accordion` core component
 
 ### Changed